requirements.txt
yfinance
pandas
newsapi-python
openai
streamlit
